tests/wpt/web-platform-tests/scroll-animations/css/scroll-timeline-shorthand.tentative.html

<!DOCTYPE html>
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/6674">
<link rel="help" href="https://drafts.csswg.org/scroll-animations-1/rewrite#scroll-timeline-axis">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script src="/css/support/shorthand-testcommon.js"></script>
<div id="target"></div>
<script>
test_valid_value('scroll-timeline', 'none block', 'none');
test_valid_value('scroll-timeline', 'none inline');
test_valid_value('scroll-timeline', 'abc horizontal');
test_valid_value('scroll-timeline', 'abc inline');
test_valid_value('scroll-timeline', 'aBc inline');
test_valid_value('scroll-timeline', 'none vertical');
test_valid_value('scroll-timeline', 'inline inline');
test_valid_value('scroll-timeline', 'abc');
test_valid_value('scroll-timeline', 'inline');
test_valid_value('view-timeline', 'a, b, c');
test_valid_value('view-timeline', 'a inline, b block, c vertical', 'a inline, b, c vertical');

test_invalid_value('scroll-timeline', '');
test_invalid_value('scroll-timeline', 'abc abc');
test_invalid_value('view-timeline', ',');
test_invalid_value('view-timeline', ',,block,,');

test_computed_value('scroll-timeline', 'none block', 'none');
test_computed_value('scroll-timeline', 'abc inline');
test_computed_value('scroll-timeline', 'none vertical', 'none vertical');
test_computed_value('scroll-timeline', 'abc horizontal');
test_computed_value('scroll-timeline', 'vertical vertical');
test_computed_value('scroll-timeline', 'abc');
test_computed_value('view-timeline', 'a, b, c');
test_computed_value('view-timeline', 'a inline, b block, c vertical', 'a inline, b, c vertical');

test_shorthand_value('scroll-timeline', 'abc vertical',
{
  'scroll-timeline-name': 'abc',
  'scroll-timeline-axis': 'vertical',
});
test_shorthand_value('scroll-timeline', 'inline horizontal',
{
  'scroll-timeline-name': 'inline',
  'scroll-timeline-axis': 'horizontal',
});
test_shorthand_value('scroll-timeline', 'abc vertical, def',
{
  'scroll-timeline-name': 'abc, def',
  'scroll-timeline-axis': 'vertical, block',
});
test_shorthand_value('scroll-timeline', 'abc, def',
{
  'scroll-timeline-name': 'abc, def',
  'scroll-timeline-axis': 'block, block',
});
</script>
